Refactor Attribute::Default to Default and Base
